% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_marked_PN.R
\name{create_marked_PN}
\alias{create_marked_PN}
\title{Create marked Petri Net}
\usage{
create_marked_PN(PN, initial_marking, final_marking)
}
\arguments{
\item{x}{a model (`list` object) returned by [pm4py_inductive()].}
}
\description{
Function to create a petri net by specifying a set of places, transitions and flows.
}
\examples{

create_PN("place_1",
			"transition_1",
			data.frame(from = "place_1",to = "transition_1"),
			marking = "place_1")

}
