% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enabled.R
\name{enabled}
\alias{enabled}
\title{Enabled transitions}
\usage{
enabled(PN)
}
\arguments{
\item{PN}{A Petri Net}
}
\description{
List the enabled transitions in a marked Petri Net. Silent transitions, i.e. starting with "inv_" or "tau" are assumed to be able to fire silently, thereby possible enabling other transitions.
}
