% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enabled_transition.R
\name{enabled_transition}
\alias{enabled_transition}
\title{Enabled Transition}
\usage{
enabled_transition(PN, transition)
}
\arguments{
\item{PN}{A Petri Net}

\item{transition}{A Transition}
}
\description{
Check if a transition is currently enabled
}
