% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute}
\alias{execute}
\title{Execute}
\usage{
execute(PN, transition)
}
\arguments{
\item{PN}{A Petri Net}

\item{transition}{The transition to be fired}
}
\description{
{Executes (fire) an enabled transition and returns the Petri Net with the New marking. If the transition is enabled
 via the firing of silent transition (i.e. starting with "inv_" of "tau"), it will fire these first. If the transition is not
 enabled, it will return FALSE.}
}
