% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_marking.R
\name{final_marking}
\alias{final_marking}
\title{Final Marking}
\usage{
final_marking(PN)
}
\arguments{
\item{PN}{A \code{\link{marked_petrinet}}}
}
\description{
Get the final marking of a \code{\link{marked_petrinet}}
}
