% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_marking.R
\name{initial_marking}
\alias{initial_marking}
\title{Initial Marking}
\usage{
initial_marking(PN)
}
\arguments{
\item{PN}{A \code{\link{marked_petrinet}}}
}
\description{
Get the initial marking of a \code{\link{marked_petrinet}}
}
