% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_nodes.R
\name{is_node}
\alias{is_node}
\title{Is node}
\usage{
is_node(node, PN)
}
\arguments{
\item{node}{\code{\link{character}} of length one: the node id to check.}

\item{PN}{\code{\link{petrinet}} or \code{\link{marked_petrinet}}}
}
\value{
logical that indicates whether \code{node} is a node in \code{PN}
}
\description{
Check if a node is part of a petri net
}
