% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_place.R
\name{is_place}
\alias{is_place}
\title{Is place}
\usage{
is_place(place, PN)
}
\arguments{
\item{place}{\code{\link{character}} of length one: the place id to check.}

\item{PN}{\code{\link{petrinet}} or \code{\link{marked_petrinet}}}
}
\value{
logical that indicates whether \code{place} is a place in \code{PN}
}
\description{
Check if a place is part of a petri net.
}
