% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_transition.R
\name{is_transition}
\alias{is_transition}
\title{Is transition}
\usage{
is_transition(transition, PN)
}
\arguments{
\item{transition}{A transition}

\item{PN}{A Petri Net}
}
\description{
Check if a transition is part of a petri net
}

