% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{nodes}
\alias{nodes}
\title{Get nodes from (marked) petrinet}
\usage{
nodes(PN)
}
\arguments{
\item{PN}{\code{\link{petrinet}} or \code{\link{marked_petrinet}}}
}
\description{
Get nodes from (marked) petrinet
}
