% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\title{Parse}
\usage{
parse(PN, trace)
}
\arguments{
\item{PN}{A Petri Net}

\item{trace}{A sequence of transitions, stored in a vector.}
}
\description{
Parses a sequence of transitions. If possible returns the Petri Net with the updated marking. Otherwise returns FALSE
}

