% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_trace.R
\name{parse_trace}
\alias{parse_trace}
\title{Parse}
\usage{
parse_trace(PN, trace)
}
\arguments{
\item{PN}{A Petri Net}

\item{trace}{A sequence of transitions, stored in a vector.}
}
\description{
Parses a sequence of transitions. If possible returns the Petri Net with the updated marking. Otherwise returns FALSE
}
