% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsel_trace.R
\name{parsel_trace}
\alias{parsel_trace}
\title{Parse (logical)}
\usage{
parsel_trace(PN, trace)
}
\arguments{
\item{PN}{A Petri Net}

\item{trace}{A sequence of transitions, stored in a vector.}
}
\description{
Tests whether a sequence of transitions can be fired by a Petri Net. If so returns TRUE, otherwise FALSE.
}
