% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petrinetR.R
\docType{package}
\name{petrinetR}
\alias{petrinetR}
\title{petrinetR - Building, visualizing, exporting and replaying Petri Nets}
\description{
Functions for the construction of Petri Nets. Petri Nets can be replayed by firing enabled transitions.
Silent transitions will be hidden by the execution handler. Also includes functionalities for the visualization of Petri Nets and
export of Petri Nets to PNML-files.
}
