% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{places}
\alias{places}
\alias{places.petrinet}
\alias{places.marked_petrinet}
\title{Places}
\usage{
places(PN)

\method{places}{petrinet}(PN)

\method{places}{marked_petrinet}(PN)
}
\arguments{
\item{PN}{A Petri Net}
}
\description{
Extracts the places from a Petri Net
}
\section{Methods (by class)}{
\itemize{
\item \code{places(petrinet)}: Places of petrinet

\item \code{places(marked_petrinet)}: Places of marked petrinet

}}
