% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postset.R
\name{post_set}
\alias{post_set}
\title{Postset}
\usage{
post_set(PN, node)
}
\arguments{
\item{PN}{A Petri Net}

\item{node}{A place or transition in the petri net}
}
\description{
Get the postset of a transition or place in a Petri Net
}

