% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset.R
\name{pre_set}
\alias{pre_set}
\title{Preset}
\usage{
pre_set(PN, node)
}
\arguments{
\item{PN}{A Petri Net}

\item{node}{A place or transition in the petri net}
}
\description{
Get the preset of a transition or place in a Petri Net
}
