% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.petrinet.R
\name{print.petrinet}
\alias{print.petrinet}
\title{Generic print function for petrinet}
\usage{
\method{print}{petrinet}(x)
}
\arguments{
\item{x}{petrinet object}

\item{...}{Additional Arguments}
}
\description{
Generic print function for petrinet
}

