% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_PN.R
\name{read_PN}
\alias{read_PN}
\title{Read PNML}
\usage{
read_PN(file, add_final_marking = TRUE)
}
\arguments{
\item{file}{Path to .pnml file}

\item{add_final_marking}{[TRUE] add final marking}
}
\description{
Function
}
