% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_PN.R
\name{render_PN}
\alias{render_PN}
\title{Render Petri Net}
\usage{
render_PN(PN)
}
\arguments{
\item{PN}{\code{\link{petrinet}} or \code{\link{marked_petrinet}}}
}
\description{
Visualize Petri Net with bipartite graph.
}
