% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_places}
\alias{n_places}
\alias{n_transitions}
\alias{n_flows}
\alias{n_nodes}
\alias{rename_transitions}
\alias{rename_places}
\alias{add_places}
\alias{add_transitions}
\alias{add_flows}
\title{Utils}
\usage{
n_places(PN)

n_transitions(PN)

n_flows(PN)

n_nodes(PN)

rename_transitions(PN, .f, ...)

rename_places(PN, .f, ...)

add_places(PN, .p)

add_transitions(PN, .t)

add_flows(PN, .flows)
}
\arguments{
\item{PN}{A petri net}

\item{.f}{A function name to apply for renaming}

\item{...}{Additional arguments}

\item{.p}{A character vector of places}

\item{.t}{A character vector of transitions}

\item{.flows}{A data.frame with a to and from column}
}
\description{
Several auxilliary functions for Petri Net objects.
}
