% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetwork_from_PN.R
\name{visNetwork_from_PN}
\alias{visNetwork_from_PN}
\title{VisNetwork from PN}
\usage{
visNetwork_from_PN(PN)
}
\arguments{
\item{PN}{Petri Net to visualize}
}
\description{
Visualize a Petri Net with an interactive network
}

