% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_marked_PN.R
\name{create_marked_PN}
\alias{create_marked_PN}
\title{Create marked Petri Net}
\usage{
create_marked_PN(PN, initial_marking, final_marking)
}
\arguments{
\item{PN}{\code{\link{petrinet}}: Object of class \code{\link{petrinet}}.}

\item{initial_marking}{\code{\link{character}}: A vector with place ids representing the initial marking.}

\item{final_marking}{\code{\link{character}}: A vector with place ids representing the final marking.}
}
\value{
A \code{\link{marked_petrinet}}
}
\description{
Function to create a  \code{\link{marked_petrinet}}, consisting of a \code{\link{petrinet}}, an initial marking, and a final marking.
}
