% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_PN.R
\name{create_PN}
\alias{create_PN}
\title{Create Petri Net}
\usage{
create_PN(places, transitions, flows)
}
\arguments{
\item{places}{A vector of unique places.}

\item{transitions}{A data.frame of unique transitions, with columns named "id" and "label"}

\item{flows}{A data.frame of flows, with columns named "from" and "to".}

\item{marking}{The names of the places to be marked.}
}
\description{
Function to create a petri net by specificing a set of places, transitions, flows and a marking.
}
\examples{

create_PN("place_1",
			"transition_1",
			data.frame(from = "place_1",to = "transition_1"),
			marking = "place_1")

}
