% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_transition.R
\name{is_transition}
\alias{is_transition}
\title{Is transition}
\usage{
is_transition(transition, PN)
}
\arguments{
\item{transition}{\code{\link{character}} of length one: the transition id to check.}

\item{PN}{\code{\link{petrinet}} or \code{\link{marked_petrinet}}}
}
\value{
logical that indicates whether \code{transition} is a transition in \code{PN}
}
\description{
Check if a transition is part of a petri net.
}
