% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset.R
\name{pre_set}
\alias{pre_set}
\title{Preset}
\usage{
pre_set(PN, node)
}
\arguments{
\item{PN}{\code{\link{petrinet}} or \code{\link{marked_petrinet}}}

\item{node}{\code{\link{character}} of length one: the node id for which to get the postset.}
}
\description{
Get the preset of a transition or place in a Petri Net
}
