% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_places}
\alias{n_places}
\alias{n_transitions}
\alias{n_flows}
\alias{n_nodes}
\alias{rename_transitions}
\alias{rename_places}
\alias{add_places}
\alias{add_transitions}
\alias{add_flows}
\title{Utils}
\usage{
n_places(PN)

n_transitions(PN)

n_flows(PN)

n_nodes(PN)

rename_transitions(PN, .f, ...)

rename_places(PN, .f, ...)

add_places(PN, places)

add_transitions(PN, transitions)

add_flows(PN, flows)
}
\arguments{
\item{PN}{A petri net}

\item{.f}{A function name to apply for renaming}

\item{...}{Additional arguments}

\item{places}{\code{data.frame} or \code{tibble} of places, with columns id and label. Both columns should be characters.}

\item{transitions}{\code{data.frame} or \code{tibble} of transitions, with columns id and label. Both columns should be characters.}

\item{flows}{\code{data.frame} or \code{tibble} of flows, with columns named "from" and "to", referring to ids of places and transitions. Both columns should be characters.}
}
\description{
Several auxilliary functions for Petri Net objects.
}
